Refactor some of the tedious logic up into ChatUtils for finding dms that include user etc
