+++ src/stream-chat-client/chatUtils.js
@@ -1,13 +1,10 @@
 import { isString, get } from 'lodash';
 
+import StreamChatClient from './client';
 import ChatRoles from './chatRoles';
 
-function getStreamUser(user) {
-  return {
-    id: user.id.split(':')[1],
-    name: `${user.profile.firstName} ${user.profile.lastName}`,
-    image: get(user, 'profile.photo.uri', ''),
-  };
+function channelIncludesUser(channel, userId) {
+  return Object.keys(get(channel, 'state.members', {})).includes(userId);
 }
 
 function getRoleFromMembership(channel) {
@@ -24,6 +21,38 @@ function getRoleFromMembership(channel) {
   return ChatRoles[role.toUpperCase()];
 }
 
+function getStreamUser(user) {
+  return {
+    id: user.id.split(':')[1],
+    name: `${user.profile.firstName} ${user.profile.lastName}`,
+    image: get(user, 'profile.photo.uri', ''),
+  };
+}
+
+async function getUserDirectMessageChannels() {
+  const currentUserId = get(StreamChatClient, 'user.id');
+  console.log('[rkd] getUserDirectMessageChannels() currentUserId:', currentUserId);
+
+  if (currentUserId) {
+    const filter = {
+      type: 'messaging',
+      members: { $in: [currentUserId] },
+    };
+    const sort = { last_message_at: -1 };
+    const options = { limit: 30 };
+
+    const dmChannelsResponse = await StreamChatClient.queryChannels(
+      filter,
+      sort,
+      options
+    );
+    console.log('[rkd] ---> dmChannelsResponse:', dmChannelsResponse);
+    return dmChannelsResponse;
+  }
+
+  return [];
+}
+
 function stripPrefix(string) {
   if (!isString) {
     return string;
@@ -33,7 +62,9 @@ function stripPrefix(string) {
 }
 
 export default {
-  getStreamUser,
+  channelIncludesUser,
   getRoleFromMembership,
+  getStreamUser,
+  getUserDirectMessageChannels,
   stripPrefix,
 };
